isl_scheduler.c: merge_edge: stop passing type
There is no need to pass in the type separately since it is already
stored in edge2.  It is not clear why the type was passed along
separately in 2a66550 (isl_schedule_constraints: add support for
conditional validity constraints, Sat Mar 24 12:09:09 2012 +0100),
since this information was available in edge2 even back then.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>