isl_scheduler.c: make detect_ccs more generally useful
In particular, take a function defining the edges as input
instead of deciding which function to use based on a "weak"
property.

This will allow us to reuse this function to detect strongly
connected components in modified graphs.  In particular, it
will allow us to keep into account the clustering when we
incrementally combine strongly connected components into
clusters.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>