isl_tab_pip.c: context_gbr_restore: abort on failure to restore cgbr->tab
If anything goes wrong during the context computation, then cgbr->tab
is set to NULL.  In particular, if cgbr->tab is NULL, then cgbr->shifted
and/or cgbr->cone may be in an inconsistent state.
Do not try perform a rollback on these tableaux in a possibly inconsistent
state.  There would not be any point in performing these rollbacks anyway.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>