isl_ast_build_node_from_schedule: generate disjunctions outside isolated part
When an isolate option is being used, the parts of the schedule domain
outside of the isolated part may be described using a disjunction.
This disjunction then gets introduced into the executed relation and
may end up getting generated at the leaves.
Although there is some support for hoisting disjunctions up, this support
is fairly limited.  Moreover, if there are any extension nodes
in between the isolation and the leaves, then the disjunction
may not apply to the extension such that it cannot even be hoisted up.

Generate the disjunction directly at the point of the isolation.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>