isl_ast_codegen.c: build_ast_from_extension: only include simple hull of domain
If the current generated domain contains any disjunctions, then
intersecting the extension with the actual generated domain
may needlessly introduce disjunctions which can complicate
the generated code.
In particular, this will happen when we introduce the disjunction
implied by the pieces of the schedule domain that lie outside
of some isolated part in the generated domain.

Intersect the extension with the simple hull of the generated domain
instead.  Note that in exceptional cases, this may result in
the generation of a subtree that does not execute any code.
If this turns out to be a problem, then we may need to introduce
additional tests to detect such situations beforehand.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>