isl_ast_graft_list_extract_shared_enforced: allow empty list of grafts
We usually do not start generating a for node unless there are
any statement instances that need to be executed by that for loop.
It therefore makes sense to error out if it turns out such
a situation occurs.

However, it is perfectly possible to recover from such situations and
so we might as well do that, even though the result will look somewhat
awkward as it will contain an empty loop guarded by an "if (0)".

Also, when we change build_ast_from_extension to only intersect
the extension with the simple hull of generated domain in a subsequent
commit, we may introduce such cases if (part of) the extension
conflicts with the generated domain, but not with its simple hull.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>