isl_basic_map_simplify: remove integer coefficients from integer divisions
If a variable has an integer coefficient inside the definition
of an integer division, then it can safely be pulled out of
the integer division.  Do so to simplify the expression.

It would be nice if we could perform this check only when there
is some potential of finding integer coefficients, but unfortunately
there are currently many places where an integer division is
introduced or modified.
Introducing integer divisions mainly happens inside set_div_from_lower_bound,
but there are several places where affine expressions are plugged
into div expressions, which may also result in integer coefficients
in the result.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>