isl_qpolynomial_fold_fold_on_domain: do not consider zero to dominate NaN
Calling isl_pw_qpolynomial_bound on { [m,n] -> -m * n } would
return max(0) because the domain is split into orthants and on
some of the orthants, the upper bound is 0, while on others
a NaN is derived.  isl_qpolynomial_fold_fold_on_domain could then
consider the NaN to be dominated by the zero, because it looks
at the sign of the difference, which is undefined for a NaN.
Ideally, the upper bound should be infinity on those orthants,
but we should at least make sure the result is not zero.

Reported-by: Fabrice RASTELLO <fabrice.rastello@inria.fr>
Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>