isl_qpolynomial_fold_fold_on_domain: return infty on unbounded input
When computing a bound over an infinite domain, a missing bound
is turned into positive or negative infinity and this value would
simply be plugged into the input polynomial, typically resulting
in NaN because the substitution engine is unaware of the signs
of the other variables and because it cannot tell if these
infinite values may cancel out.

Detect the situations where this may happen and produce
a +infty or -infty result instead.  This result may be wildly
inaccurate, but it is always a correct bound and it should
be no worse than returning NaN.

Requested-by: Fabrice RASTELLO <fabrice.rastello@inria.fr>
Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>