isl_ast_build_node_from_schedule: improve handling of isolation
In particular, improve the detection of the case where we are
in a branch outside the isolation domain at an outer level
(such that there is no need to further split the schedule domain).

If the AST generator was able to eliminate a schedule dimension
in this outside branch, then it has also been eliminated
from the domain of the executed relation.  No conflict can
therefore be detected with the constraints on this dimension
in the isolated domain unless we also plug in the value
we found for the eliminated dimension in this domain.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>