rename isl_schedule_*_foreach_* to isl_schedule_*_foreach_*_top_down
It is not clear from the original name whether the tree is traversed
top down or bottom up, which may be confusing because
isl_schedule_*map_* traverse the tree in the other direction.
Furthermore, we may want to add a variant of foreach that traverses
the tree in the other direction.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>