Make pass-by-reference explicit by using pointers
GMP's mpz_t automatically decays to pointer and IMath's mp_int is a
pointer, therefore changing them in a function body always also changes
the argument of the function's caller. With an isl_int implementation
without pointer semantics this wouldn't be the case and the value
change not visible outside of the function.

The parameters a, b and g of isl_int_gcdext are output-parameters, so
their return values must be visible to the function's caller. We change
them to pointers to make the by-reference explicit.

Signed-off-by: Michael Kruse <isl@meinersbur.de>
Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>