replace --schedule-fuse option by --schedule-serialize-sccs option
The only effect of setting the --schedule-fuse option to min
was to serialize the strongly connected components in the dependence
graph as soon as they arise.  It is better to let the option name
reflect this effect more directly since the --schedule-fuse option
is not about fusion per se.  In particular, the default value "max"
in itself does not result in maximal fusion and may therefore be
confusing to users.

Internally, the schedule_fuse option is completely removed.
It is also removed from the C API.  However, the command line
option is kept for now to ease the transition.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>