Revert "scheduler: also (optionally) bound coefficients computed by carry_dependences"
This reverts commit 095dbe323b6e6ae3ad92c210eb012723d7a8438c.

It makes little sense to try and bound the coefficients while looking
for a rational solution since we cannot prevent the LP solver from
increasing the denominator of the solution, allowing arbitrarily
large numerators.  It is these numerators that are used to construct
the corresponding schedule row.

Moreover, adding these extra constraints breaks assumptions
about what the solution of the LP problem will look like.
In particular, the bounded dependence distances could attain
values that are different from the expected extremes of zero and one.
Their sum may therefore not be an integer either,
leading to an inappropriate invocation of the isl_int_divexact
clearing the denominator of the presumed integer value
and resulting in undefined behavior.
While it is possible to avoid making these assumptions,
it is not worth it to do it for something that makes little sense.

Add a test case that could break before.

Reported-by: Tobias Grosser <tobias@grosser.es>
Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>