isl_map_simplify.c: uset_gist_full: negate constraint in copy
While constructing a test set for checking whether a constraint is redundant,
negate the constraint that is to be tested in the copy rather than
first dropping the constraint by pretending it is redundant and then
adding the opposite.

This should make the code easier to follow and it makes it easier
to change the meaning of bset and combined in the next commit.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>