isl_map_simplify.c uset_gist_full: detect equalities in the tableau
Since uset_gist currently performs a compression based on the equalities
in the intersection of the input and the context, there can not be
any equalities left, so there is no point in checking for equalities now.
However, in a future commit, uset_gist will only take into account
some equalities for the compression and then the tableau may indeed
contain hidden equalities.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>