isl_map_simplify.c: uset_gist_full: allow equalities in context
Since uset_gist currently performs a compression based on the equalities
in the intersection of the input and the context, there can not be
any equalities left in the compressed context.
However, in a future commit, uset_gist will only take into account
some equalities for the compression and then the compressed context
may involve equalities.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>