isl_basic_map_gist: only take into account equalities in input and context
In particular, do not take into account additional equalities
in the intersection of input and context for computing a compression.
While taking into account such additional equalities can result
in extra equalities getting detected in the gist, it also makes
it impossible or at least difficult to avoid introducing constraints
in the gist on variables that do not appear in the input.

Introducing such constraints is not only counterintuitive,
it also breaks an assumption in the AST generator that
the result of the gist will not involve any loop iterators
that did not already appear in the input.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>