isl_map_simplify.c: update_ineq: take into account tableau may be empty
Currently, uset_gist computes the affine hull of the intersection
of the input with the context and this affine hull is tested for emptiness.
This means that the tableau constructed for identifying redundant constraints
can never be empty.  However, in a future commit, this intersection
will no longer be computed.  update_ineq therefore needs to take
into account that the tableau may be empty.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>