isl_map_simplify.c: uset_gist: drop redundant intersection
Now that the affine hull used in the compression is the intersection
of the affine hulls of input and context (rather than the affine hull
of the intersection of input and context), there is no longer any
need to intersect the input with this affine hull.
The affine hull of the input is still available in the input
now that uset_gist_full preserves the original constraints in the input,
while the affine hull of the context is taken into account
while reducing the constraints using those equalities.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>