isl_map_simplify.c: uset_gist: do not introduce variables that were not involved
Introducing variables in the gist that did not appear in the input
is not only counterintuitive,
it also breaks an assumption in the AST generator that
the result of the gist will not involve any loop iterators
that did not already appear in the input.

isl_basic_map_gist (the caller of uset_gist) may still introduce
such variables, but this will be taken care of in the next commit.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>