isl_basic_map_gist: do not introduce context divs in input
The divs of the input and the context are currently aligned,
which introduces the context divs in the input.  This may seem
harmless because there should not be any constraints involving
those context divs in the result such that those divs will get
eliminated again, but this only happens after the div constraints
have been added and they may get simplified in between.
This may lead to spurious constraints in the output that may
even involve variables that did not appear in the input.

Do not add the context divs to the input, but add completely
unconstrained dimensions instead.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>