add -std=gnu99 to compile flags when compiling imath-32 backend with gcc
The imath-32 backend relies on the C99 behavior of "inline", which is
different from the default gcc behavior, resulting in multiple definition
errors when compiling the imath-32 backend with gcc.
The original imath-32 support was developed on top of a patch
that would end up setting -std=gnu99 for gcc, but this patch was dropped.
Set -std=gnu99 explicitly when compiling the imath-32 backend with gcc.

Reported-by: Tobias Grosser <tobias@grosser.es>
Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>