isl_union_*_plain_is_equal: also check total number of parts
isl_union_*_plain_is_equal already checks the number of elements
in the hash table, which is currently equal to the number of parts.
However, in a subsequent commit, the representation will be changed
from a single level hash table to a two level hash table, in which
case the number of parts may no longer be equal to the number of
elements in the hash table.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>