isl_union_*_add_part_generic: check for overlapping domains instead of spaces
A UNION is currently not allowed to contain multiple entries with
the same domain space, but this will be changed in a subsequent commit.
In that case and if "disjoint" is set, isl_union_*_add_part_generic needs
to check that the domains of two entries that have the same domain space
are disjoint.  Introduce the infrastructure for performing these tests,
which are currently specialized to the case where no two entries can have
the same domain space.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>