isl_ast_build_set_loop_bounds: extract out update of pending and generated
This allows the caller to perform some operations between the detection
of whether the loop is degenerate or eliminated and the update
of the pending and generated sets.  In particular, it will allow
the effect of the next commit to be minimized.
The specialization of the bounds also needs to be extracted out
since the specialized version is used both in
isl_ast_build_set_loop_bounds and in the new
isl_ast_build_set_pending_generated.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>