isl_basic_map_output_defining_equality: also detect modulo expressions
A modulo expression of the form { [i] -> [i mod 4] } is represented
internally as

    { [i] -> [o] : 4*floor((-i + o)/4) = -i + o and 0 <= o <= 3 }

isl_basic_map_output_defining_equality would not recognize
the equality 4*floor((-i + o)/4) = -i + o as defining o in terms
of the input dimensions, because the equality involves an integer
division that references output dimensions.
This reference can however be eliminated by taking into account
the constraints o >= 0 and o <= 3.

Extend isl_basic_map_output_defining_equality to also detect this pattern and
extend plain_pw_multi_aff_from_map to extract an affine expression
in case this pattern is detected.

The effect of this change is visible on the output of the unroll4 test case.

The slightly negative effect on the p.delft2-0 is due to the fact that
before, an expression involving floor((-1 + 4P2 - 2c3/9)
would be derived, whereas now the expression involves floor((-4P2 + 2c3/9).
In the first case, extract_nonneg_mod checks the non-negativity
of -1 + 4P2 - 2c3 and 9 - 4P2 + 2c3 and the second one succeeds.
In the second case, extract_nonneg_mod checks the non-negativity
of -4P2 + 2c3 and 8 + 4P2 - 2c3, neither of which succeeds.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>