isl_ast_codegen.c: generate_domain: extract out add_domain
In an upcoming commit, the gisting of the inverse schedule with respect
with the build domain will be cut short if the inverse schedule is
obviously single valued.  The operations that need to be performed
before or after the gisting are the same, so extract them into
a separate function first.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>