isl_ast_codegen.c: add_domain: specialize after gist
add_domain is currently (usually) called on an inverse schedule that has
already been gisted.  In the next commit, this gisting will be skipped
in more cases, which may result in spurious extra constraints in the guard.
In particular, an extra "n == 0" condition would be generated
in the component6 test case.
The guard is already gisted inside add_domain, but this is performed
after calling isl_ast_build_specialize, which may reduce gisting
opportunities.  Do it beforehand instead.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>