isl_map_simplify.c: uset_gist_compressed: take into account context may be empty
Ever since c6d0c8c (isl_basic_map_gist: only take into account equalities
in input and context, Sat Jun 20 18:34:41 2015 +0200), the compression
performed on input and context is based on the intersection of
the affine hulls of input and context rather than the affine hull
of the intersection.  This means that the context may become empty
as a result of the compression in case the input and the context
do not intersect.  This situation is not handled by c85043a
(isl_map_simplify.c: update_ineq: take into account tableau may be empty,
Sat Jun 20 18:34:41 2015 +0200) because isl_tab_from_basic_set
treats obviously empty basic sets in a special way, breaking
assumptions in uset_gist_full about the positions of the inequalities.

Reported-by: Tobias Grosser <tobias@grosser.es>
Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>